requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests
uamqp
azure-storage-queue==2.1.0
avro==1.10.0
azure-common==1.1.28
azure-core==1.23.1
azure-eventhub==5.7.0
azure-identity==1.9.0
azure-schemaregistry==1.0.0
azure-schemaregistry-avroserializer==1.0.0b4
azure-storage-common==2.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
idna==3.3
isodate==0.6.1
msal==1.17.0
msal-extensions==0.3.1
msrest==0.6.21
oauthlib==3.2.0
portalocker==2.4.0
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.1.1
urllib3==1.26.9
